feat: unify upload mapping workflow
