+++ apps/react-ui/client/src/components/DataPreview/index.ts
@@ -1,0 +1,2 @@
+export { default as DataPreview } from "./DataPreview";
+export type { DataPreviewProps } from "./DataPreview";
